Invokes and Debug
