+++ src/NeoSharp.Application/Controllers/PromptInvokesController.cs
@@ -1,0 +1,45 @@
+﻿using System;
+using NeoSharp.Application.Attributes;
+using NeoSharp.Application.Client;
+using NeoSharp.Core.Models;
+using NeoSharp.Core.Types;
+
+namespace NeoSharp.Application.Controllers
+{
+    public class PromptInvokesController : IPromptController
+    {
+        #region Private fields
+
+        private readonly IConsoleWriter _consoleWriter;
+        private readonly IConsoleReader _consoleReader;
+
+        #endregion
+
+        /// <summary>
+        /// Constructor
+        /// </summary>
+        /// <param name="consoleWriter">Console writter</param>
+        /// <param name="consoleReader">Console reader</param>
+        public PromptInvokesController(IConsoleWriter consoleWriter, IConsoleReader consoleReader)
+        {
+            _consoleReader = consoleReader;
+            _consoleWriter = consoleWriter;
+        }
+
+        // testinvoke {contract hash} {params} (--attach-neo={amount}, --attach-gas={amount}) (--from-addr={addr})
+
+        /// <summary>
+        /// Invoke contract
+        /// </summary>
+        /// <param name="contractHash">Contract</param>
+        /// <param name="body">Body</param>
+        [PromptCommand("invoke", Help = "Invoke a contract", Category = "Invokes")]
+        public void Invoke(UInt160 contractHash, [PromptCommandParameterBody] object[] args)
+        {
+            Contract contract = Contract.GetContract(contractHash);
+            if (contract == null) throw (new ArgumentNullException("Contract not found"));
+
+            var tx = contract.CreateInvokeTransaction(args);
+        }
+    }
+}
